# swagger.yml
openapi: 3.0.0
info:
  title: School Management System
  description: This is a School Management System which provides automation of various academic and administrative operations.
  version: 1.0.0
  contact:
    email: ankuragarwa246@gmail.com
servers:
  - url: http://localhost:9999/api
    description: Development server
  - url: https://sch-management-server.herokuapp.com/api
    description: Production Server
paths:
...
  /student/fetchProfile:
    get:
      tags:
        - Student
      summary: Fetch Student Profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Student Profile fetched Successfully
        '401':
          description: Not Authorized to perform this action
        '404':
          description: Student Not Found
        '500':
          description: Internal Server Error
  /student/updateProfile:
    put:
      tags:
        - Student
      summary: Update Student Profile
      security:
        - bearerAuth: []
      requestBody:
        $ref: '#/components/requestBodies/StudentProfile'
      responses:
        '200':
          description: Student Profile Updated Successfully
        '400':
          description: Invalid input
        '401':
          description: Not Authorized to perform this action
        '500':
          description: Internal Server Error
  /student/scorecards:
    get:
      tags:
        - Student
      summary: Fetch all the scorecards of the student
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Scorecards Fetched Successfully
        '401':
          description: Not Authorized to perform this action
        '500':
          description: Internal Server Error
  /library/addBook:
    post:
      tags:
        - Library
      summary: Add a book to the library
      security:
        - bearerAuth: []
      requestBody:
        $ref: '#/components/requestBodies/Book'
      responses:
        '201':
          description: Book Added Successfully
        '400':
          description: Invalid input
        '401':
          description: Not Authorized to perform this action
        '500':
          description: Internal Server Error
  /library/updateBook/{bookId}:
    put:
      tags:
        - Library
      summary: Update a book in the library
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: bookId
          required: true
          schema:
            type: string
      requestBody:
        $ref: '#/components/requestBodies/Book'
      responses:
        '200':
          description: Book Updated Successfully
        '400':
          description: Invalid input
        '401':
          description: Not Authorized to perform this action
        '404':
          description: Book Not Found
        '500':
          description: Internal Server Error
  /library/deleteBook/{bookId}:
    delete:
      tags:
        - Library
      summary: Delete a book from the library
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: bookId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Book Deleted Successfully
        '400':
          description: Book is issued to a student, cannot be deleted
        '401':
          description: Not Authorized to perform this action
        '404':
          description: Book Not Found
        '500':
          description: Internal Server Error
  /library/getLateStudents:
    get:
      tags:
        - Library
      summary: Fetch all the students who haven't returned the books in time
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Students Fetched Successfully
        '401':
          description: Not Authorized to perform this action
        '500':
          description: Internal Server Error
  /library/issueBook/{bookId}/student/{studentId}:
    post:
      tags:
        - Library
      summary: Issue a book to a student
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: bookId
          required: true
          schema:
            type: string
        - in: path
          name: studentId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Book Issued Successfully
        '400':
          description: Invalid input
        '401':
          description: Not Authorized to perform this action
        '404':
          description: Book Not Found
        '500':
          description: Internal Server Error
  /library/returnBook/{bookId}/student/{studentId}:
    post:
      tags:
        - Library
      summary: Return a book to library
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: bookId
          required: true
          schema:
            type: string
        - in: path
          name: studentId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Book Returned Successfully
        '400':
          description: Invalid input
        '401':
          description: Not Authorized to perform this action
        '404':
          description: Book Not Found
        '500':
          description: Internal Server Error
components:
  schemas:
...
    StudentProfileInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        dob:
          type: string
          format: date-time
        gender:
          type: string
...
    BookInput:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        quantity:
          type: number
        price:
          type: number
  securitySchemes:
    bearerAuth:
      name: Authorization
      type: http
      scheme: bearer
      in: header
      bearerFormat: JWT
  requestBodies:
...
    StudentProfile:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StudentProfileInput'
        description: Student Profile
...
    Book:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BookInput'
        description: Book Creation